Refactor CodeIgniter class
Add Bootstrap class and test for CodeIgniter class
