+++ tests/_support/MockCodeIgniter.php
@@ -1,0 +1,8 @@
+<?php namespace CodeIgniter;
+
+use Config\Services;
+
+class MockCodeIgniter extends CodeIgniter
+{
+
+}
